sadatom: allow fractional per-l occupations in --occs
--occs parsed into std::vector<int> and stored Eigen::VectorXi, so only
integral per-l counts could be pinned. That prevents scanning E(n)
along a fractional 4s/3d split, which is the direct way to locate the
minimum and measure the curvature of the energy along the occupation
coordinate.

The downstream consumer already works in real arithmetic: scf.cpp
casts each entry to OOO_Real and hands it to OOO's
fixed_number_of_particles_per_block, which takes reals. So only the
storage type and the parser change; scf.cpp needs no edit.

The Hund auto-split branch (lmax+1 totals for an unrestricted run)
distributes electrons over n-shells with integer arithmetic and has no
meaningful fractional generalisation, so it keeps requiring integers
and now says so explicitly rather than silently truncating.

twodquadrature.cpp assigns the tabulated PBE ground-state config to
fixed_per_l_a; Eigen has no implicit scalar conversion, so that
assignment needs an explicit .cast<double>().

Verified: the integer Hund-split form is unchanged (--occs="8 12 6"
runs as before); the guard fires on --occs="4 6 5.5"; and fractional
counts are genuinely applied rather than rounded, shown by the energy
varying smoothly along the occupation coordinate. Scanning Fe M=5
lmax=2 with alpha=(4,6,5), beta=(3+x,6,2-x) over x=0.50..0.80 in PBE
gives a clean parabola (max residual 4.5e-6 Ha) with the minimum at
x=0.65404, i.e. 3d_beta=1.34596, and d2E/dx2=0.0666 Ha. That curvature
means a +-0.0035 displacement in occupation costs only 4.1e-7 Ha,
which is why chained restarts of the free-occupation run scatter over
3d_beta=1.336-1.343 while the energy barely moves.

+++ src/sadatom/scf.h
@@ -63,8 +63,11 @@ namespace helfem {
         // 2*(2l+1)) and leave fixed_per_l_b empty.
         // Unrestricted: pass both; each entry is the alpha or beta
         // count in that l (up to 2l+1).
-        Eigen::VectorXi fixed_per_l_a;
-        Eigen::VectorXi fixed_per_l_b;
+        // Counts may be fractional: OOO's frozen-occupation API takes
+        // reals, so a partially filled l shell can be pinned at any
+        // occupation in [0, 2l+1] (or [0, 2*(2l+1)] when restricted).
+        Eigen::VectorXd fixed_per_l_a;
+        Eigen::VectorXd fixed_per_l_b;
         /// OOO verbosity; 0 for silent, higher for per-iteration prints.
         int verbosity = 5;
         /// SCF convergence algorithms handed to OOO's state machine: a
